requirements.txt
streamlit
pillow
opencv-python
matplotlib